working on C++ file organization for nimbleLists
