+++ packages/nimble/configure.win
@@ -4,7 +4,7 @@
 cp inst/CppCode/dists.cpp src
 cp inst/CppCode/RcppUtils.cpp src
 cp inst/CppCode/nimbleGraph.cpp src
-cp inst/CppCode/EigenTypedefs.cpp src
+cp inst/CppCode/eigenUsingClasses.cpp src
