+++ packages/nimble/inst/CppCode/RcppNimbleUtils.cpp
@@ -1175,11 +1175,3 @@ void nimble_optim_withVarArgs(void* nimFun, OptimControl* control, OptimAns* ans
 }	
 
 
-SEXP makeNewNimbleList(SEXP S_listName) {
-  SEXP call;
-  PROTECT(call = allocVector(LANGSXP, 2));
-  SETCAR(call, install("makeNewNimListSEXPRESSIONFromC"));
-  SETCADR(call, S_listName);
-  UNPROTECT(1);
-  return(EVAL(call));
-}
